requirements.txt
cloudpickle==2.2.1
kaggle==1.5.13
Pillow==9.4.0
Pillow==9.5.0
protobuf==3.20.3
qdrant_client==1.1.6
sentence_transformers==2.2.2
streamlit==1.21.0
toml==0.10.2
tqdm==4.65.0
google-cloud-storage==2.8.0